requirements.txt
asgiref==3.4.1
Django==2.2.10
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.8.6
gunicorn==20.1.0